refactor sensor config loading
Agisoft does not always write only the actually used sensors to the xml
file, but sometimes also additional sensors, which previously caused
problems during load. This change creates placeholder data for invalid
configurations. When the `intrinsics_mat` with is consructed later,
the available sensors are filtered and an error is only raised, when
an invalid sensor config is going to be used.

Additionally, the configurations are now stored in a map instead of a
list, so the order and continuity of sensor ids in the file does not
matter anymore.

Also, the global width and height of images is set to the values of the
last used sensor, instead of the first sensor in the file to avoid
trouble with oddly formatted files.