First fix for portrait images
if the image is in landscape modus, the current implementation does not
respect the changed dimensions and tries to read the depth map out of
bounds. this change swaps the x and y coordinates if the image height is
larger than the image width.

should work but needs a real fix with information from the agisoft xml
that is already available in the scene